refactor(dht): move behaviours into own respective module
