+++ market_dht/src/request.rs
@@ -1,5 +1,4 @@
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub(crate) enum Request {}
-
 pub(crate) enum RequestOk {}
 pub(crate) enum RequestError {}
